Add release workflow to be used by balto-utils
We created the release-tagger reusable workflow in #7 but did not add it
to balto-utils itself.

There is a _chance_ we could do this by adding a second trigger to the
release-tagger workflow (with the push tags filter), but I worry that it
could be confusing to others (or even future me) as the tags filter
would have to be ignored for external usage.

After this is merged, I'll delete the v1 branch