Simplify yarn & npm cache
Currently, we cache the entirety of node_modules. The setup-node action
has caching built in, but it explicitly _does not_ cache node_modules
and instead caches the global caches of npm/yarn/etc. The stated
reasoning from their docs is that it allows you to maintain
compatibility between different node versions.

After adding support for ESLint v9 in balto-eslint, we saw an issue
where linting was failing due to a module resolution issue. After using
a branch with these changes, those issues went away.

In this commit, we're switching from our own managed version of caching
node_modules to leaning into the built-in caching from setup-node.
Performance wise, it's _about_ the same when looking at the entire run.
The initial install is slower, but saving the cache in the post-run step
is faster.

Overall, it feels like a win to reduce complexity and let setup manage
the caching. As a bonus, there are currently warnings about using an
outdated version of actions/cache which will be going away with this
change.